Moved build stuff to new build.py, and cleaned up
