+++ pythonforandroid/recipes/freetype/__init__.py
@@ -1,5 +1,5 @@
 
-from pythonforandroid.toolchain import Recipe, shprint, get_directory, current_directory, ArchARM
+from pythonforandroid.toolchain import Recipe, shprint, current_directory, ArchARM
 from os.path import exists, join, realpath
 from os import uname
 import glob
