+++ pythonforandroid/util.py
@@ -125,3 +125,20 @@ def which(program, path_env):
                 return exe_file
 
     return None
+
+
+def get_directory(filename):
+    '''If the filename ends with a recognised file extension, return the
+    filename without this extension.'''
+    if filename.endswith('.tar.gz'):
+        return basename(filename[:-7])
+    elif filename.endswith('.tgz'):
+        return basename(filename[:-4])
+    elif filename.endswith('.tar.bz2'):
+        return basename(filename[:-8])
+    elif filename.endswith('.tbz2'):
+        return basename(filename[:-5])
+    elif filename.endswith('.zip'):
+        return basename(filename[:-4])
+    info('Unknown file extension for {}'.format(filename))
+    exit(1)
